#3044 refactoring and minor fixes
